requirements.txt
# Minerva AI Assistant - Required Dependencies

# Core dependencies
loguru>=0.6.0
click>=8.1.3
rich>=12.5.1
python-dotenv>=0.21.0
pydantic>=1.10.2

# Web interface (with fixed versions for WebSocket compatibility)
Flask==2.2.5
Werkzeug==2.2.3
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
eventlet>=0.33.3
Flask-WTF>=1.1.1

# AI and Machine Learning
openai>=1.0.0
langchain>=0.0.267
langchain_openai>=0.0.1
transformers>=4.28.1
huggingface_hub>=0.14.1
diffusers>=0.14.0
accelerate>=0.19.0
torch>=2.0.0
sentence-transformers>=2.2.2
bitsandbytes>=0.41.0  # For 8-bit quantization support
scipy>=1.10.0  # Required by many ML libraries
einops>=0.6.1  # Required by some advanced models

# Multi-model integration
gpt4all>=1.0.0  # GPT4All integration
psutil>=5.9.0   # System monitoring
asyncio>=3.4.3  # For asynchronous model processing
aiohttp>=3.8.1  # For async HTTP requests

# Voice interface 
speechrecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13
whisper>=1.1.10

# Framework integrations
aiohttp>=3.8.5
requests>=2.31.0
websockets>=11.0.3

# Memory and storage
chromadb>=0.4.18
sqlalchemy>=2.0.23
alembic>=1.12.1
redis>=4.6.0

# API and service integrations
fastapi>=0.104.1
uvicorn>=0.24.0
websocket-client>=1.6.4
pyjwt>=2.8.0

# Utilities
numpy>=1.24.3
pandas>=2.0.0
tqdm>=4.66.1
pillow>=10.0.0
python-multipart>=0.0.6

# Testing
pytest>=7.3.1
pytest-flask>=1.2.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-socket>=0.7.0
markdown>=3.4.3
